# Only provide fully qualified Component completion if unimported.