--- a/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/IsExternalInit.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/IsExternalInit.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace System.Runtime.CompilerServices
+{
+    // Used to compile against C# 9 in a netstandard2.0 app.
+    internal class IsExternalInit
+    {
+    }
+}