--- a/src/Razor/test/Microsoft.CodeAnalysis.Razor.Workspaces.Test/DefaultTagHelperCompletionServiceTest.cs
+++ b/src/Razor/test/Microsoft.CodeAnalysis.Razor.Workspaces.Test/DefaultTagHelperCompletionServiceTest.cs
@@ -634,6 +634,44 @@ namespace Microsoft.VisualStudio.Editor.Razor
         }
 
         [Fact]
+        public void GetElementCompletions_FiltersFullyQualifiedElementsIfShortNameExists()
+        {
+            // Arrange
+            var documentDescriptors = new[]
+            {
+                TagHelperDescriptorBuilder.Create("TestTagHelper", "TestAssembly")
+                    .TagMatchingRuleDescriptor(rule => rule.RequireTagName("Test"))
+                    .Build(),
+                TagHelperDescriptorBuilder.Create("TestTagHelper", "TestAssembly")
+                    .TagMatchingRuleDescriptor(rule => rule.RequireTagName("TestAssembly.Test"))
+                    .AddMetadata(ComponentMetadata.Component.NameMatchKey, ComponentMetadata.Component.FullyQualifiedNameMatch)
+                    .Build(),
+                TagHelperDescriptorBuilder.Create("Test2TagHelper", "TestAssembly")
+                    .TagMatchingRuleDescriptor(rule => rule.RequireTagName("Test2Assembly.Test"))
+                    .AddMetadata(ComponentMetadata.Component.NameMatchKey, ComponentMetadata.Component.FullyQualifiedNameMatch)
+                    .Build(),
+            };
+            var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
+            {
+                ["Test"] = new HashSet<TagHelperDescriptor>() { documentDescriptors[0] },
+                ["Test2Assembly.Test"] = new HashSet<TagHelperDescriptor>() { documentDescriptors[2] },
+            });
+
+            var completionContext = BuildElementCompletionContext(
+                documentDescriptors,
+                Array.Empty<string>(),
+                containingTagName: "body",
+                containingParentTagName: null);
+            var service = CreateTagHelperCompletionFactsService();
+
+            // Act
+            var completions = service.GetElementCompletions(completionContext);
+
+            // Assert
+            AssertCompletionsAreEquivalent(expectedCompletions, completions);
+        }
+
+        [Fact]
         public void GetElementCompletions_TagOutputHintDoesNotFallThroughToSchemaCheck()
         {
             // Arrange