- The intent of this change is to improve performance so our payloads are "less"
- Prior to this we'd always provide component completions for every listed component twice over. Meaning, if you had a `<Counter>` you'd get a `Counter` completion and a `MyNamespace.Counter` completion item. Now we remove the fully qualified variants if their non-fully qualified variant is in scope.
- Added a test to validate the new completion behavior.

**Before:**
|                                              Method |      Mean |     Error |    StdDev |   Op/s |    Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|---------:|---------:|--------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 11.219 ms | 0.0953 ms | 0.0744 ms |  89.13 | 718.7500 | 359.3750 |       - |  4,569 KB |
|           'Component Completion List Serialization' |  2.221 ms | 0.0442 ms | 0.1008 ms | 450.33 |  82.0313 |  27.3438 | 15.6250 |    555 KB |
|         'Component Completion List Deserialization' |  8.801 ms | 0.0836 ms | 0.0698 ms | 113.63 | 640.6250 | 312.5000 |       - |  4,013 KB |

**After:**
|                                              Method |     Mean |     Error |    StdDev |  Op/s |    Gen 0 |    Gen 1 | Gen 2 | Allocated |
|---------------------------------------------------- |---------:|----------:|----------:|------:|---------:|---------:|------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 5.195 ms | 0.0605 ms | 0.0472 ms | 192.5 | 367.1875 | 179.6875 |     - |  2,294 KB |
|           'Component Completion List Serialization' | 1.071 ms | 0.0212 ms | 0.0236 ms | 934.0 |  44.9219 |   7.8125 |     - |    287 KB |
|         'Component Completion List Deserialization' | 4.127 ms | 0.0748 ms | 0.1049 ms | 242.3 | 320.3125 | 148.4375 |     - |  2,007 KB |

**Results:**
Roundtrip: 2.2x faster
Serialization: 2.1x faster
Deserialization: 2.5x faster

Fixes dotnet/aspnetcore#31939